update handle queue binding
